Refactors sign-up related code into it's own component.
